Total re-do of the database setup/init
Will no longer rely on migration files in the Sequelize-based code, to
set up the database. Schema is now in db/, along with CSV files for the
initial data and scripts for creating the CSV files and for creating
the DB itself.

The package.json file in express-server is updated so that the "reset"
task handles the new actions.

All files in express-server/src/seeders are removed. The files in
express-server/src/migrations can go, too, at some point. Right now,
keeping them for reference.